refactor: organize usecases folders
